Moved the MediaManager widget from the CMS module to the System module
