+++ modules/backend/widgets/mediamanager/assets/js/mediamanager-browser.js
