+++ modules/backend/widgets/mediamanager/assets/js/mediamanager-global.js
