+++ modules/backend/widgets/mediamanager/assets/js/mediamanager.imagecroppopup.js
