+++ modules/backend/widgets/mediamanager/assets/js/mediamanager.js
