+++ modules/backend/widgets/mediamanager/assets/js/mediamanager.popup.js
