+++ modules/backend/widgets/mediamanager/assets/less/mediamanager.less
