+++ modules/backend/widgets/mediamanager/partials/_bottom-toolbar.htm
