+++ modules/backend/widgets/mediamanager/partials/_crop-tool-image-area.htm
