+++ modules/backend/widgets/mediamanager/partials/_crop-toolbar.htm
