+++ modules/backend/widgets/mediamanager/partials/_generic-list.htm
