+++ modules/backend/widgets/mediamanager/partials/_item-icon.htm
