+++ modules/backend/widgets/mediamanager/partials/_list-grid.htm
