+++ modules/backend/widgets/mediamanager/partials/_move-form.htm
