+++ modules/backend/widgets/mediamanager/partials/_new-folder-form.htm
