+++ modules/backend/widgets/mediamanager/partials/_popup-body.htm
