+++ modules/backend/widgets/mediamanager/partials/_rename-form.htm
