+++ modules/backend/widgets/mediamanager/partials/_resize-image-form.htm
