+++ modules/backend/widgets/mediamanager/partials/_thumbnail-image.htm
