+++ modules/backend/widgets/mediamanager/partials/_upload-progress.htm
