+++ modules/backend/widgets/mediamanager/partials/_view-mode-buttons.htm
